Extract sun session UI
